+++ client_src/js/lib/functions/ui.js
@@ -139,4 +139,9 @@ export default class UI {
       return false;
     }
   }
+  static async delay(ms) {
+    return new Promise((resolve) => {
+      setTimeout(resolve, ms);
+    });
+  }
 }
