+++ client_src/js/printer-management.runner.js
@@ -7,1112 +7,91 @@ import { updatePrinterSettingsModal } from "./lib/modules/printerSettings.js";
 import FileOperations from "./lib/functions/file.js";
 import Validate from "./lib/functions/validate.js";
 import PowerButton from "./lib/modules/powerButton.js";
-import {
-  checkQuickConnectState,
-  init as actionButtonInit
-} from "./lib/modules/Printers/actionButtons.js";
 import PrinterSelect from "./lib/modules/printerSelect";
-import PrinterLogs from "./lib/modules/printerLogs.js";
 import CustomGenerator from "./lib/modules/customScripts.js";
 import Sortable from "./vendor/sortable";
 
 import { updateConnectionLog } from "./printerManager/connection-log";
 import { createOrUpdatePrinterTableRow } from "./printerManager/printer-data.js";
+import {
+  bulkOctoPrintClientUpdate,
+  bulkOctoPrintPluginUpdate,
+  bulkConnectPrinters,
+  bulkDisconnectPrinters,
+  bulkOctoPrintPowerCommand,
+  bulkOctoPrintPreHeatCommand,
+  bulkOctoPrintControlCommand
+} from "./printerManager/functions/bulk-commands-functions";
+import { scanNetworkForDevices } from "./printerManager/functions/printer-manager.functions";
 
 const workerURL = "/printersInfo/get/";
 
-let printerInfo = "";
 const deletedPrinters = [];
 let powerTimer = 5000;
 
-let newPrintersIndex = 0;
-const removeLine = function (element) {
-  element.remove();
-};
-// Dash control listeners
-let bulkPluginUpdateButton = document.getElementById("blkUpdatePluginsBtn");
-bulkPluginUpdateButton.addEventListener("click", async (e) => {
-  let currentPrinterList = await OctoFarmClient.post("printers/printerInfo");
-  currentPrinterList = await currentPrinterList.json();
-  let message = "";
-  let toUpdate = [];
-  let pluginList = [];
-  for (let printer = 0; printer < currentPrinterList.length; printer++) {
-    let currentPrinter = currentPrinterList[printer];
-    if (currentPrinter.octoPrintPluginUpdates.length > 0) {
-      message += currentPrinter.printerName + "<br>";
-      toUpdate.push({
-        printerURL: currentPrinter.printerURL,
-        printerName: currentPrinter.printerName,
-        apikey: currentPrinter.apikey
-      });
-      for (
-        let plugin = 0;
-        plugin < currentPrinter.octoPrintPluginUpdates.length;
-        plugin++
-      ) {
-        let currentPlugin = currentPrinter.octoPrintPluginUpdates[plugin];
-        pluginList.push(currentPlugin.id);
-      }
-    }
-  }
+const multiPrinterSelectModal = document.getElementById("multiPrintersSection");
 
-  message += "Are you sure?";
-  bootbox.confirm({
-    size: "medium",
-    title: "This will update the following printers plugins...",
-    message: message,
-    callback: async function (result) {
-      if (result) {
-        for (let i = 0; i < toUpdate.length; i++) {
-          const data = {
-            targets: pluginList,
-            force: true
-          };
-          let updateRequest = await OctoPrintClient.postNOAPI(
-            toUpdate[i],
-            "plugin/softwareupdate/update",
-            data
-          );
-          if (updateRequest.status === 200) {
-            UI.createAlert(
-              "success",
-              `${toUpdate[i].printerName}: Successfully updated! your instance will restart now.`,
-              3000,
-              "Clicked"
-            );
-            let post = await OctoPrintClient.systemNoConfirm(
-              toUpdate[i],
-              "restart"
-            );
-            if (typeof post !== "undefined") {
-              if (post.status === 204) {
-                UI.createAlert(
-                  "success",
-                  `Successfully made restart attempt to ${toUpdate[i].printerName}... You may need to Re-Sync!`,
-                  3000,
-                  "Clicked"
-                );
-              } else {
-                UI.createAlert(
-                  "error",
-                  `There was an issue sending restart to ${toUpdate[i].printerName} are you sure it's online?`,
-                  3000,
-                  "Clicked"
-                );
-              }
-            } else {
-              UI.createAlert(
-                "error",
-                `No response from ${toUpdate[i].printerName}, is it online???`,
-                3000,
-                "Clicked"
-              );
-            }
-          } else {
-            UI.createAlert(
-              "error",
-              `${toUpdate[i].printerName}: Failed to update, manual intervention required!`,
-              3000,
-              "Clicked"
-            );
-          }
-        }
-      }
-    }
-  });
+// Bulk OctoPrint Command Listeners
+let bulkPluginUpdateButton = document.getElementById("blkUpdatePluginsBtn");
+bulkPluginUpdateButton.addEventListener("click", async () => {
+  await bulkOctoPrintPluginUpdate();
 });
 
 let bulkOctoPrintUpdateButton = document.getElementById("blkOctoPrintUpdate");
 bulkOctoPrintUpdateButton.addEventListener("click", async (e) => {
-  let currentPrinterList = await OctoFarmClient.post("printers/printerInfo");
-  currentPrinterList = await currentPrinterList.json();
-  let message = "";
-  let toUpdate = [];
-  for (let printer = 0; printer < currentPrinterList.length; printer++) {
-    let currentPrinter = currentPrinterList[printer];
-    if (currentPrinter.octoPrintUpdate.updateAvailable) {
-      message += currentPrinter.printerName + "<br>";
-
-      toUpdate.push({
-        printerURL: currentPrinter.printerURL,
-        printerName: currentPrinter.printerName,
-        apikey: currentPrinter.apikey
-      });
-    }
-  }
-
-  message += "Are you sure?";
-  bootbox.confirm({
-    size: "medium",
-    title: "This will update the following OctoPrint Installs...",
-    message: message,
-    callback: async function (result) {
-      if (result) {
-        for (let i = 0; i < toUpdate.length; i++) {
-          const data = {
-            targets: ["octoprint"],
-            force: true
-          };
-          let updateRequest = await OctoPrintClient.postNOAPI(
-            toUpdate[i],
-            "plugin/softwareupdate/update",
-            data
-          );
-          if (updateRequest.status === 200) {
-            UI.createAlert(
-              "success",
-              `${toUpdate[i].printerName}: Update command fired, you may need to restart OctoPrint once complete.`,
-              3000,
-              "Clicked"
-            );
-          } else {
-            UI.createAlert(
-              "error",
-              `${toUpdate[i].printerName}: Failed to update, manual intervention required!`,
-              3000,
-              "Clicked"
-            );
-          }
-        }
-      }
-    }
-  });
-});
-
-const deleteAllBtn = document.getElementById("delAllBtn");
-
-async function delay(ms) {
-  return new Promise((resolve) => {
-    setTimeout(resolve, ms);
-  });
-}
-
-deleteAllBtn.addEventListener("click", async (e) => {
-  let onScreenButtons = document.querySelectorAll("*[id^=delButton-]");
-  for (const btn of onScreenButtons) {
-    btn.click();
-  }
-});
-const saveAllBtn = document.getElementById("saveAllBtn");
-saveAllBtn.addEventListener("click", async (e) => {
-  saveAllBtn.disabled = true;
-  deleteAllBtn.disabled = true;
-  let onScreenButtons = document.querySelectorAll("*[id^=saveButton-]");
-  let onScreenDelete = document.querySelectorAll("*[id^=delButton-]");
-  for (const btn of onScreenButtons) {
-    btn.disabled = true;
-  }
-  for (const btn of onScreenDelete) {
-    btn.disabled = true;
-  }
-  UI.createAlert(
-    "warning",
-    "Starting to save all your instances... this may take some time...",
-    onScreenButtons.length * 1500
-  );
-  for (const btn of onScreenButtons) {
-    btn.disabled = false;
-    btn.click();
-    await delay(1500);
-  }
-  UI.createAlert("success", "Successfully saved all your instances", 4000);
-  saveAllBtn.disabled = false;
-  deleteAllBtn.disabled = true;
+  await bulkOctoPrintClientUpdate();
 });
 
 const bulkConnectBtn = document.getElementById("bulkConnectBtn");
 bulkConnectBtn.addEventListener("click", async (e) => {
-  const printerConnect = async function () {
-    let printersToConnect = [];
-    //Grab all check boxes
-    const selectedPrinters = PrinterSelect.getSelected();
-    selectedPrinters.forEach((element) => {
-      const ca = element.id.split("-");
-      printersToConnect.push(ca[1]);
-    });
-    for (let p = 0; p < printersToConnect.length; p++) {
-      const index = _.findIndex(printerInfo, function (o) {
-        return o._id === printersToConnect[p];
-      });
-      if (index > -1) {
-        let data = {};
-        if (typeof printerInfo[index].connectionOptions !== "undefined") {
-          data = {
-            command: "connect",
-            port: printerInfo[index].connectionOptions.portPreference,
-            baudrate: printerInfo[index].connectionOptions.baudratePreference,
-            printerProfile:
-              printerInfo[index].connectionOptions.printerProfilePreference,
-            save: true
-          };
-        } else {
-          UI.createAlert(
-            "warning",
-            `${printerInfo[index].printerName} has no preferences saved, defaulting to AUTO...`,
-            8000,
-            "Clicked"
-          );
-          data.command = "connect";
-          data.port = "AUTO";
-          data.baudrate = 0;
-          data.printerProfile = "_default";
-          data.save = false;
-        }
-        if (
-          printerInfo[index].printerState.colour.category === "Disconnected"
-        ) {
-          let post = await OctoPrintClient.post(
-            printerInfo[index],
-            "connection",
-            data
-          );
-          if (typeof post !== "undefined") {
-            if (post.status === 204) {
-              UI.createAlert(
-                "success",
-                `Successfully made connection attempt to ${printerInfo[index].printerName}...`,
-                3000,
-                "Clicked"
-              );
-            } else {
-              UI.createAlert(
-                "error",
-                `There was an issue connecting to ${printerInfo[index].printerName} it's either not online, or the connection options supplied are not available...`,
-                3000,
-                "Clicked"
-              );
-            }
-          } else {
-            UI.createAlert(
-              "error",
-              `No response from ${printerInfo[index].printerName}, is it online???`,
-              3000,
-              "Clicked"
-            );
-          }
-        } else {
-          UI.createAlert(
-            "warning",
-            `Printer ${printerInfo[index].printerName} is not in "Disconnected" state... skipping`,
-            3000,
-            "Clicked"
-          );
-        }
-      } else {
-        UI.createAlert(
-          "error",
-          `Could not find ${printerInfo[index].printerName} in your printer in the list of available printers...`,
-          3000,
-          "Clicked"
-        );
-      }
-    }
-  };
-
-  PrinterSelect.create(
-    document.getElementById("multiPrintersSection"),
+  await PrinterSelect.create(
+    multiPrinterSelectModal,
     false,
     "Connect Printers",
-    printerConnect
+    bulkConnectPrinters
   );
 });
 const bulkDisconnectBtn = document.getElementById("bulkDisconnectBtn");
 bulkDisconnectBtn.addEventListener("click", async (e) => {
-  const printerDisconnect = async function () {
-    let printersToConnect = [];
-    //Grab all check boxes
-    const selectedPrinters = PrinterSelect.getSelected();
-    selectedPrinters.forEach((element) => {
-      const ca = element.id.split("-");
-      printersToConnect.push(ca[1]);
-    });
-    for (let p = 0; p < printersToConnect.length; p++) {
-      const index = _.findIndex(printerInfo, function (o) {
-        return o._id === printersToConnect[p];
-      });
-      if (index > -1) {
-        let data = {
-          command: "disconnect"
-        };
-        if (printerInfo[index].printerState.colour.category === "Idle") {
-          let post = await OctoPrintClient.post(
-            printerInfo[index],
-            "connection",
-            data
-          );
-          if (typeof post !== "undefined") {
-            if (post.status === 204) {
-              UI.createAlert(
-                "success",
-                `Successfully made disconnect attempt to ${printerInfo[index].printerName}...`,
-                3000,
-                "Clicked"
-              );
-            } else {
-              UI.createAlert(
-                "error",
-                `There was an issue disconnecting to ${printerInfo[index].printerName} are you sure it's online?`,
-                3000,
-                "Clicked"
-              );
-            }
-          } else {
-            UI.createAlert(
-              "error",
-              `No response from ${printerInfo[index].printerName}, is it online???`,
-              3000,
-              "Clicked"
-            );
-          }
-        } else {
-          UI.createAlert(
-            "warning",
-            `Printer ${printerInfo[index].printerName} is not in "Idle" state... skipping`,
-            3000,
-            "Clicked"
-          );
-        }
-      } else {
-        UI.createAlert(
-          "error",
-          `Could not find ${printerInfo[index].printerName} in your printer in the list of available printers...`,
-          3000,
-          "Clicked"
-        );
-      }
-    }
-  };
-
-  PrinterSelect.create(
-    document.getElementById("multiPrintersSection"),
+  await PrinterSelect.create(
+    multiPrinterSelectModal,
     false,
     "Disconnect Printers",
-    printerDisconnect
+    bulkDisconnectPrinters
   );
 });
 const bulkPowerBtn = document.getElementById("bulkPowerBtn");
 bulkPowerBtn.addEventListener("click", async (e) => {
-  const printerPower = async function () {
-    bootbox.prompt({
-      title: "Power command!",
-      message: "<p>Please select an option below:</p>",
-      inputType: "radio",
-      inputOptions: [
-        {
-          text: "Restart OctoPrint",
-          value: "restart"
-        },
-        {
-          text: "Reboot Host",
-          value: "reboot"
-        },
-        {
-          text: "Shutdown Host",
-          value: "shutdown"
-        }
-      ],
-      callback: async function (result) {
-        let printersToConnect = [];
-        //Grab all check boxes
-        const selectedPrinters = PrinterSelect.getSelected();
-        selectedPrinters.forEach((element) => {
-          const ca = element.id.split("-");
-          printersToConnect.push(ca[1]);
-        });
-        for (let p = 0; p < printersToConnect.length; p++) {
-          const index = _.findIndex(printerInfo, function (o) {
-            return o._id === printersToConnect[p];
-          });
-          if (index > -1) {
-            if (printerInfo[index].printerState.colour.category !== "Active") {
-              let post = await OctoPrintClient.systemNoConfirm(
-                printerInfo[index],
-                result
-              );
-              await delay(1000);
-              if (typeof post !== "undefined") {
-                if (post.status === 204) {
-                  UI.createAlert(
-                    "success",
-                    `Successfully made ${result} attempt to ${printerInfo[index].printerName}...`,
-                    3000,
-                    "Clicked"
-                  );
-                } else {
-                  UI.createAlert(
-                    "error",
-                    `There was an issue sending ${result} to ${printerInfo[index].printerName} are you sure it's online?`,
-                    3000,
-                    "Clicked"
-                  );
-                }
-              } else {
-                UI.createAlert(
-                  "error",
-                  `No response from ${printerInfo[index].printerName}, is it online???`,
-                  3000,
-                  "Clicked"
-                );
-              }
-            } else {
-              UI.createAlert(
-                "warning",
-                `Printer ${printerInfo[index].printerName} is not in "Idle" state... skipping`,
-                3000,
-                "Clicked"
-              );
-            }
-          } else {
-            UI.createAlert(
-              "error",
-              "Could not find your printer in your printer in the list of available printers...",
-              3000,
-              "Clicked"
-            );
-          }
-        }
-      }
-    });
-  };
-
-  PrinterSelect.create(
+  await PrinterSelect.create(
     document.getElementById("multiPrintersSection"),
     false,
     "Power On/Off Printers",
-    printerPower
+    bulkOctoPrintPowerCommand
   );
 });
 
 let scanNetworkBtn = document.getElementById("scanNetworkBtn");
 scanNetworkBtn.addEventListener("click", async (e) => {
-  e.target.disabled = true;
-  UI.createAlert(
-    "warning",
-    "Scanning your network for new devices now... Please wait!",
-    20000
-  );
-  let printers = await OctoFarmClient.get("printers/scanNetwork");
-  let scannedPrinters = await printers.json();
-  for (let index = 0; index < scannedPrinters.length; index++) {
-    const printer = {
-      printerURL: "",
-      cameraURL: "",
-      name: "",
-      group: "",
-      apikey: ""
-    };
-
-    if (typeof scannedPrinters[index].name !== "undefined") {
-      printer.name = scannedPrinters[index].name;
-    }
-    if (typeof scannedPrinters[index].url !== "undefined") {
-      printer.printerURL = scannedPrinters[index].url;
-    }
-    PrintersManagement.addPrinter(printer);
-  }
-  UI.createAlert(
-    "success",
-    "Scanning your network for new devices now... Please wait!",
-    3000,
-    "Clicked"
-  );
-  e.target.disabled = false;
+  await scanNetworkForDevices();
 });
 
 let bulkPreHeat = document.getElementById("bulkPreHeat");
 bulkPreHeat.addEventListener("click", async (e) => {
-  const printersPreHeat = async function () {
-    let printersToConnect = [];
-    //Grab all check boxes
-    const selectedPrinters = PrinterSelect.getSelected();
-    selectedPrinters.forEach((element) => {
-      const ca = element.id.split("-");
-      printersToConnect.push(ca[1]);
-    });
-
-    bootbox.dialog({
-      title: "Bulk printer heating...",
-      message: `
-        <form class="form-inline">
-           <div class="input-group mb-3">
-            <div class="input-group-prepend">
-              <label class="input-group-text" for="preHeatToolSelect">Tool #: </label>
-            </div>
-            <select class="custom-select" id="preHeatToolSelect">
-              <option selected value="0">0</option>
-              <option value="0">1</option>
-              <option value="0">2</option>
-              <option value="0">3</option>
-              <option value="0">4</option>
-            </select>
-          </div>
-          <div class="input-group mb-3">
-            <input type="text" class="form-control" placeholder="0" aria-label="0" aria-describedby="basic-addon1" id="preHeatToolTempSelect">
-            <div class="input-group-append">
-              <span class="input-group-text" id="basic-addon1">°C</span>
-            </div>
-          </div>
-          <p>&nbsp;</p>
-          <div class="input-group mb-3">
-            <div class="input-group-prepend">
-              <span class="input-group-text" id="basic-addon1">Bed</span>
-            </div>
-            <input type="text" class="form-control" placeholder="0" aria-label="0" aria-describedby="basic-addon1" id="preHeatBedTempSelect">
-            <div class="input-group-append">
-              <span class="input-group-text" id="basic-addon1">°C</span>
-            </div>
-          </div>
-          <p>&nbsp;</p>
-          <div class="input-group mb-3">
-            <div class="input-group-prepend">
-              <span class="input-group-text" id="basic-addon1">Chamber</span>
-            </div>
-            <input type="text" class="form-control" placeholder="0" aria-label="0" aria-describedby="basic-addon1" id="preHeatChamberTempSelect">
-                  <div class="input-group-append">
-              <span class="input-group-text" id="basic-addon1">°C</span>
-            </div>
-          </div>
-        </form>
-        `,
-      size: "large",
-      onEscape: true,
-      backdrop: true,
-      closeButton: true,
-      buttons: {
-        action: {
-          label: '<i class="fas fa-fire"></i> Heat!',
-          className: "btn-success",
-          callback: async function () {
-            let toolNumber = document.getElementById("preHeatToolSelect");
-            let toolTemp = document.getElementById("preHeatToolTempSelect");
-            let bedTemp = document.getElementById("preHeatBedTempSelect");
-            let chamberTemp = document.getElementById(
-              "preHeatChamberTempSelect"
-            );
-            let toolData = {
-              command: "target",
-              targets: {}
-            };
-            let bedData = {
-              command: "target",
-              target: 0
-            };
-            let chamberData = {
-              command: "target",
-              target: 0
-            };
-
-            if (toolTemp.value !== "" && !isNaN(toolTemp.value)) {
-              toolData.targets["tool" + toolNumber.value] = parseInt(
-                toolTemp.value
-              );
-            }
-            if (bedTemp.value !== "" && !isNaN(bedTemp.value)) {
-              bedData.target = parseInt(bedTemp.value);
-            }
-            if (chamberTemp.value !== "" && !isNaN(chamberTemp.value)) {
-              chamberData.target = parseInt(chamberTemp.value);
-            }
-            for (let p = 0; p < printersToConnect.length; p++) {
-              const index = _.findIndex(printerInfo, function (o) {
-                return o._id === printersToConnect[p];
-              });
-              if (index > -1) {
-                //Set tool temp
-                if (toolTemp.value !== "" && !isNaN(toolTemp.value)) {
-                  let post = await OctoPrintClient.post(
-                    printerInfo[index],
-                    "printer/tool",
-                    toolData
-                  );
-                  if (typeof post !== "undefined") {
-                    if (post.status === 204) {
-                      UI.createAlert(
-                        "success",
-                        `Successfully set tool${toolNumber.value} target attempt to ${printerInfo[index].printerName}...`,
-                        3000,
-                        "Clicked"
-                      );
-                    } else {
-                      UI.createAlert(
-                        "error",
-                        `There was an issue setting tool${toolNumber.value} target attempt to ${printerInfo[index].printerName} are you sure it's online?`,
-                        3000,
-                        "Clicked"
-                      );
-                    }
-                  } else {
-                    UI.createAlert(
-                      "error",
-                      `No response from ${printerInfo[index].printerName}, is it online???`,
-                      3000,
-                      "Clicked"
-                    );
-                  }
-                }
-                //Set bed temp
-                if (bedTemp.value !== "" && !isNaN(bedTemp.value)) {
-                  let post = await OctoPrintClient.post(
-                    printerInfo[index],
-                    "printer/bed",
-                    bedData
-                  );
-                  if (typeof post !== "undefined") {
-                    if (post.status === 204) {
-                      UI.createAlert(
-                        "success",
-                        `Successfully set bed target attempt to ${printerInfo[index].printerName}...`,
-                        3000,
-                        "Clicked"
-                      );
-                    } else {
-                      UI.createAlert(
-                        "error",
-                        `There was an issue setting bed target attempt to ${printerInfo[index].printerName} are you sure it's online?`,
-                        3000,
-                        "Clicked"
-                      );
-                    }
-                  } else {
-                    UI.createAlert(
-                      "error",
-                      `No response from ${printerInfo[index].printerName}, is it online???`,
-                      3000,
-                      "Clicked"
-                    );
-                  }
-                }
-                //Set chamber temp
-                if (chamberTemp.value !== "" && !isNaN(chamberTemp.value)) {
-                  let post = await OctoPrintClient.post(
-                    printerInfo[index],
-                    "printer/chamber",
-                    chamberData
-                  );
-                  if (typeof post !== "undefined") {
-                    if (post.status === 204) {
-                      UI.createAlert(
-                        "success",
-                        `Successfully set chamber target attempt to ${printerInfo[index].printerName}...`,
-                        3000,
-                        "Clicked"
-                      );
-                    } else {
-                      UI.createAlert(
-                        "error",
-                        `There was an issue setting chamber target attempt to ${printerInfo[index].printerName} are you sure it's online?`,
-                        3000,
-                        "Clicked"
-                      );
-                    }
-                  } else {
-                    UI.createAlert(
-                      "error",
-                      `No response from ${printerInfo[index].printerName}, is it online???`,
-                      3000,
-                      "Clicked"
-                    );
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-    });
-  };
-
   PrinterSelect.create(
-    document.getElementById("multiPrintersSection"),
+    multiPrinterSelectModal,
     false,
     "Pre-Heat Printers",
-    printersPreHeat
+    await bulkOctoPrintPreHeatCommand
   );
 });
 
 let bulkControl = document.getElementById("bulkControl");
 bulkControl.addEventListener("click", async (e) => {
-  const printersControl = async function () {
-    let printersToConnect = [];
-    //Grab all check boxes
-    const selectedPrinters = PrinterSelect.getSelected();
-    selectedPrinters.forEach((element) => {
-      const ca = element.id.split("-");
-      printersToConnect.push(ca[1]);
-    });
-    let cameraBlock = "";
-
-    printersToConnect.forEach((printer) => {
-      const index = _.findIndex(printerInfo, function (o) {
-        return o._id === printer;
-      });
-      if (index > -1) {
-        cameraBlock += `
-        <div class="col-lg-3">
-            <img width="100%" src="${printerInfo[index].cameraURL}">
-        </div>
-        `;
-      }
-    });
-
-    bootbox.dialog({
-      title: "Bulk printer control...",
-      message: `
-      <div id="printerControls" class="row">
-            <div class="col-lg-12">
-              <div class="row">
-                 ${cameraBlock}
-              </div>
-            </div>
-            <div class="col-md-6">
-                            <div class="row">
-                    <div class="col-9">
-                        <center>
-                            <h5>X/Y</h5>
-                        </center>
-                        <hr>
-                    </div>
-                    <div class="col-3">
-                        <center>
-                            <h5>Z</h5>
-                        </center>
-                        <hr>
-                    </div>
-                </div>
-                <div class="row">
-                    <div class="col-3"></div>
-                    <div class="col-3">
-                        <center><button id="pcYpos" type="button" class="btn btn-light"><i class="fas fa-arrow-up"></i></button></center>
-                    </div>
-                    <div class="col-3"></div>
-                    <div class="col-3">
-                        <center><button id="pcZpos" type="button" class="btn btn-light"><i class="fas fa-arrow-up"></i></button></center>
-                    </div>
-                </div>
-                <div class="row">
-                    <div class="col-3">
-                        <center><button id="pcXneg" type="button" class="btn btn-light"><i class="fas fa-arrow-left"></i></button></center>
-                    </div>
-                    <div class="col-3">
-                        <center><button id="pcXYhome" type="button" class="btn btn-light"><i class="fas fa-home"></i></button></center>
-                    </div>
-                    <div class="col-3">
-                        <center><button id="pcXpos" type="button" class="btn btn-light"><i class="fas fa-arrow-right"></i></button></center>
-                    </div>
-                    <div class="col-3">
-                        <center><button id="pcZhome" type="button" class="btn btn-light"><i class="fas fa-home"></i></button></center>
-                    </div>
-                </div>
-                <div class="row">
-                    <div class="col-3"></div>
-                    <div class="col-3">
-                        <center><button id="pcYneg" type="button" class="btn btn-light"><i class="fas fa-arrow-down"></i></button></center>
-                    </div>
-                    <div class="col-3"></div>
-                    <div class="col-3">
-                        <center><button id="pcZneg" type="button" class="btn btn-light"><i class="fas fa-arrow-down"></i></button></center>
-                    </div>
-                </div>
-                <div class="row">
-                    <div class="col-12">
-                        <center>
-                            <div id="pcAxisSteps" class="btn-group" role="group">
-                                <button id="pcAxisSteps01" type="button" class="btn btn-light" value="01">0.1</button>
-                                <button id="pcAxisSteps1" type="button" class="btn btn-light" value="1">1</button>
-                                <button id="pcAxisSteps10" type="button" class="btn btn-dark active" value="10">10</button>
-                                <button id="pcAxisSteps100" type="button" class="btn btn-light" value="100">100</button>
-                            </div>
-                        </center>
-                    </div>
-                </div>
-            </div>
-            <div class="col-md-6">
-            <center>
-            <button id="pmPrintStart" type="button" class="btn btn-success mb-1" role="button" style="display: inline-block;"><i class="fas fa-print"></i> Print</button> <br>
-            <button id="pmPrintPause" type="button" class="btn btn-light mb-1" role="button" style="display: inline-block;"><i class="fas fa-pause"></i> Pause</button> <br>
-            <button id="pmPrintRestart" type="button" class="btn btn-warning mb-1" role="button"  style="display: inline-block;"><i class="fas fa-undo"></i> Restart</button> <br>
-            <button id="pmPrintResume" type="button" class="btn btn-info mb-1" role="button"  style="display: inline-block;"><i class="fas fa-redo"></i> Resume</button> <br>
-            <button id="pmPrintStop" type="button" class="btn btn-danger mb-1" style="display: inline-block;"><i class="fas fa-square"></i> Cancel</button> <br>
-            </center>
-            </div>
-        `,
-      size: "large",
-      onEscape: true,
-      backdrop: true,
-      closeButton: true,
-      onShow: function (e) {
-        //Grab Page
-        const printerControls = {
-          xPlus: document.getElementById("pcXpos"),
-          xMinus: document.getElementById("pcXneg"),
-          yPlus: document.getElementById("pcYpos"),
-          yMinus: document.getElementById("pcYneg"),
-          xyHome: document.getElementById("pcXYhome"),
-          zPlus: document.getElementById("pcZpos"),
-          zMinus: document.getElementById("pcZneg"),
-          zHome: document.getElementById("pcZhome"),
-          step01: document.getElementById("pcAxisSteps01"),
-          step1: document.getElementById("pcAxisSteps1"),
-          step10: document.getElementById("pcAxisSteps10"),
-          step100: document.getElementById("pcAxisSteps100"),
-          printStart: document.getElementById("pmPrintStart"),
-          printPause: document.getElementById("pmPrintPause"),
-          printRestart: document.getElementById("pmPrintRestart"),
-          printResume: document.getElementById("pmPrintResume"),
-          printStop: document.getElementById("pmPrintStop")
-        };
-        printerControls.printStart.addEventListener("click", async (e) => {
-          printersToConnect.forEach((printer) => {
-            const index = _.findIndex(printerInfo, function (o) {
-              return o._id === printer;
-            });
-            if (index > -1) {
-              e.target.disabled = true;
-              const opts = {
-                command: "start"
-              };
-
-              OctoPrintClient.jobAction(printerInfo[index], opts, e);
-            }
-          });
-        });
-        printerControls.printPause.addEventListener("click", (e) => {
-          printersToConnect.forEach((printer) => {
-            const index = _.findIndex(printerInfo, function (o) {
-              return o._id === printer;
-            });
-            if (index > -1) {
-              e.target.disabled = true;
-              const opts = {
-                command: "pause",
-                action: "pause"
-              };
-              OctoPrintClient.jobAction(printerInfo[index], opts, e);
-            }
-          });
-        });
-        printerControls.printRestart.addEventListener("click", (e) => {
-          printersToConnect.forEach((printer) => {
-            const index = _.findIndex(printerInfo, function (o) {
-              return o._id === printer;
-            });
-            if (index > -1) {
-              e.target.disabled = true;
-              const opts = {
-                command: "restart"
-              };
-              OctoPrintClient.jobAction(printerInfo[index], opts, e);
-            }
-          });
-        });
-        printerControls.printResume.addEventListener("click", (e) => {
-          printersToConnect.forEach((printer) => {
-            const index = _.findIndex(printerInfo, function (o) {
-              return o._id === printer;
-            });
-            if (index > -1) {
-              e.target.disabled = true;
-              const opts = {
-                command: "pause",
-                action: "resume"
-              };
-              OctoPrintClient.jobAction(printerInfo[index], opts, e);
-            }
-          });
-        });
-        printerControls.printStop.addEventListener("click", (e) => {
-          bootbox.confirm({
-            message:
-              "Are you sure you want to cancel all of your ongoing print?",
-            buttons: {
-              cancel: {
-                label: '<i class="fa fa-times"></i> Cancel'
-              },
-              confirm: {
-                label: '<i class="fa fa-check"></i> Confirm'
-              }
-            },
-            callback(result) {
-              if (result) {
-                printersToConnect.forEach((printer) => {
-                  const index = _.findIndex(printerInfo, function (o) {
-                    return o._id === printer;
-                  });
-                  if (index > -1) {
-                    e.target.disabled = true;
-                    const opts = {
-                      command: "cancel"
-                    };
-                    OctoPrintClient.jobAction(printerInfo[index], opts, e);
-                  }
-                });
-              }
-            }
-          });
-        });
-
-        printerControls.xPlus.addEventListener("click", (e) => {
-          printersToConnect.forEach((printer) => {
-            const index = _.findIndex(printerInfo, function (o) {
-              return o._id === printer;
-            });
-            if (index > -1) {
-              OctoPrintClient.move(e, printerInfo[index], "jog", "x");
-            }
-          });
-        });
-        printerControls.xMinus.addEventListener("click", (e) => {
-          printersToConnect.forEach((printer) => {
-            const index = _.findIndex(printerInfo, function (o) {
-              return o._id === printer;
-            });
-            if (index > -1) {
-              OctoPrintClient.move(e, printerInfo[index], "jog", "x", "-");
-            }
-          });
-        });
-        printerControls.yPlus.addEventListener("click", (e) => {
-          printersToConnect.forEach((printer) => {
-            const index = _.findIndex(printerInfo, function (o) {
-              return o._id === printer;
-            });
-            if (index > -1) {
-              OctoPrintClient.move(e, printerInfo[index], "jog", "y");
-            }
-          });
-        });
-        printerControls.yMinus.addEventListener("click", (e) => {
-          printersToConnect.forEach((printer) => {
-            const index = _.findIndex(printerInfo, function (o) {
-              return o._id === printer;
-            });
-            if (index > -1) {
-              OctoPrintClient.move(e, printerInfo[index], "jog", "y", "-");
-            }
-          });
-        });
-        printerControls.xyHome.addEventListener("click", (e) => {
-          printersToConnect.forEach((printer) => {
-            const index = _.findIndex(printerInfo, function (o) {
-              return o._id === printer;
-            });
-            if (index > -1) {
-              OctoPrintClient.move(e, printerInfo[index], "home", ["x", "y"]);
-            }
-          });
-        });
-        printerControls.zPlus.addEventListener("click", (e) => {
-          printersToConnect.forEach((printer) => {
-            const index = _.findIndex(printerInfo, function (o) {
-              return o._id === printer;
-            });
-            if (index > -1) {
-              OctoPrintClient.move(e, printerInfo[index], "jog", "z");
-            }
-          });
-        });
-        printerControls.zMinus.addEventListener("click", (e) => {
-          printersToConnect.forEach((printer) => {
-            const index = _.findIndex(printerInfo, function (o) {
-              return o._id === printer;
-            });
-            if (index > -1) {
-              OctoPrintClient.move(e, printerInfo[index], "jog", "z", "-");
-            }
-          });
-        });
-        printerControls.zHome.addEventListener("click", (e) => {
-          printersToConnect.forEach((printer) => {
-            const index = _.findIndex(printerInfo, function (o) {
-              return o._id === printer;
-            });
-            if (index > -1) {
-              OctoPrintClient.move(e, printerInfo[index], "home", ["z"]);
-            }
-          });
-        });
-        printerControls.step01.addEventListener("click", (e) => {
-          printersToConnect.forEach((printer) => {
-            const index = _.findIndex(printerInfo, function (o) {
-              return o._id === printer;
-            });
-            if (index > -1) {
-              OctoFarmClient.post("printers/stepChange", {
-                printer: printerInfo[index]._id,
-                newSteps: "01"
-              });
-            }
-          });
-
-          printerControls.step01.className = "btn btn-dark active";
-          printerControls.step1.className = "btn btn-light";
-          printerControls.step10.className = "btn btn-light";
-          printerControls.step100.className = "btn btn-light";
-        });
-        printerControls.step1.addEventListener("click", (e) => {
-          printersToConnect.forEach((printer) => {
-            const index = _.findIndex(printerInfo, function (o) {
-              return o._id === printer;
-            });
-            if (index > -1) {
-              OctoFarmClient.post("printers/stepChange", {
-                printer: printerInfo[index]._id,
-                newSteps: "1"
-              });
-            }
-          });
-
-          printerControls.step1.className = "btn btn-dark active";
-          printerControls.step01.className = "btn btn-light";
-          printerControls.step10.className = "btn btn-light";
-          printerControls.step100.className = "btn btn-light";
-        });
-        printerControls.step10.addEventListener("click", (e) => {
-          printersToConnect.forEach((printer) => {
-            const index = _.findIndex(printerInfo, function (o) {
-              return o._id === printer;
-            });
-            if (index > -1) {
-              OctoFarmClient.post("printers/stepChange", {
-                printer: printerInfo[index]._id,
-                newSteps: "10"
-              });
-            }
-          });
-
-          printerControls.step10.className = "btn btn-dark active";
-          printerControls.step1.className = "btn btn-light";
-          printerControls.step01.className = "btn btn-light";
-          printerControls.step100.className = "btn btn-light";
-        });
-        printerControls.step100.addEventListener("click", (e) => {
-          printersToConnect.forEach((printer) => {
-            const index = _.findIndex(printerInfo, function (o) {
-              return o._id === printer;
-            });
-            if (index > -1) {
-              OctoFarmClient.post("printers/stepChange", {
-                printer: printerInfo[index]._id,
-                newSteps: "100"
-              });
-            }
-          });
-
-          printerControls.step100.className = "btn btn-dark active";
-          printerControls.step1.className = "btn btn-light";
-          printerControls.step10.className = "btn btn-light";
-          printerControls.step01.className = "btn btn-light";
-        });
-      }
-    });
-  };
-
   PrinterSelect.create(
     document.getElementById("multiPrintersSection"),
     false,
     "Control Printers",
-    printersControl
+    bulkOctoPrintControlCommand()
   );
 });
 
@@ -1390,77 +369,6 @@ function drawScriptTable(scripts) {
           .classList.toggle("d-none");
       }
     });
-}
-
-function pluginListTemplate(plugin) {
-  //Also need check inplace for incompatible...
-  let abandoned = "";
-  if (plugin.abandoned === true) {
-    abandoned =
-      '<i class="fa fa-heartbeat" title="Abandoned by its maintainer"></i>';
-  }
-  let latestRelease = "";
-  if (
-    typeof plugin.github !== "undefined" &&
-    typeof plugin.github.latest_release !== "undefined"
-  ) {
-    latestRelease = `
-    <small class="prop" title="Github stars"><i class="fa fa-star"></i> <span >${plugin.github.stars}</span></small>
-    <small class="prop" title="Last push to main branch"><i class="fa fa-refresh"></i> <span>${plugin.github.stars}</span></small>
-    <small class="prop" title="Latest release &amp; date"><i class="fa fa-tag"></i> <span >${plugin.github.latest_release.tag}</span> (<span >${plugin.github.latest_release.date}</span>)</small>`;
-  }
-
-  return `
-      <div class="entry" id="plugin-${plugin.title.replace(/ /g, "_")}">
-                        <div class="row-fluid">
-                            <div class="span12">
-                                <div>
-                                    <span>${plugin.title}</span>
-
-                                    ${abandoned}
-                           
-                                </div>
-                                <div class="meta">
-                                    <small class="prop">
-                                      <i class="fa fa-info"></i>
-                                      &nbsp;<a target="_blank" href="${
-                                        plugin.page
-                                      }" title="${plugin.page}">
-                                      Details
-                                      </a>
-                                    </small>
-                                    <small class="prop">
-                                        <i class="fa fa-home"></i>
-                                        &nbsp;<a target="_blank" href="${
-                                          plugin.homepage
-                                        }" title="${plugin.homepage}">
-                                        Homepage
-                                        </a>
-                                      </small>
-                                    <small class="prop"><i class="fa fa-user"></i> <span title="${
-                                      plugin.author
-                                    }">${plugin.author}</span></small>
-                                </div>
-                                <div class="muted"><small>${
-                                  plugin.description
-                                }</small></div>
-                                <div class="stats">
-                                    <small class="prop" title="License"><i class="fa fa-gavel"></i> <span>${
-                                      plugin.license
-                                    }</span></small>
-                                    <small class="prop" title="Publication date"><i class="fa fa-birthday-cake"></i> <span>${JSON.stringify(
-                                      plugin.published
-                                    ).substring(1, 11)}</span></small>
-                                    <small class="prop" title="Active instances past month"><i class="fa fa-server"></i> <span >${JSON.stringify(
-                                      plugin.stats.instances_month
-                                    ).substring(0, 11)}</span></small>
-                                    ${latestRelease}
-                                </div>
-                            </div>
-
-                        </div>
-                    </div>
-  `;
 }
 
 const pluginAction = async function (action) {
@@ -2197,311 +1105,41 @@ document.getElementById("addPrinterBtn").addEventListener("click", (event) => {
   PrintersManagement.addPrinter();
 });
 
-class Printer {
-  constructor(printerURL, camURL, apikey, group, name) {
-    this.settingsAppearance = {
-      color: "default",
-      colorTransparent: false,
-      defaultLanguage: "_default",
-      name,
-      showFahrenheitAlso: false
-    };
-    this.printerURL = printerURL;
-    this.camURL = camURL;
-    this.apikey = apikey;
-    this.group = group;
+const deleteAllBtn = document.getElementById("delAllBtn");
+deleteAllBtn.addEventListener("click", async (e) => {
+  let onScreenButtons = document.querySelectorAll("*[id^=delButton-]");
+  for (const btn of onScreenButtons) {
+    btn.click();
   }
-}
-
-class PrintersManagement {
-  constructor(printerURL, camURL, apikey, group, name) {
-    this.printer = new Printer(printerURL, camURL, apikey, group, name);
+});
+const saveAllBtn = document.getElementById("saveAllBtn");
+saveAllBtn.addEventListener("click", async (e) => {
+  saveAllBtn.disabled = true;
+  deleteAllBtn.disabled = true;
+  let onScreenButtons = document.querySelectorAll("*[id^=saveButton-]");
+  let onScreenDelete = document.querySelectorAll("*[id^=delButton-]");
+  for (const btn of onScreenButtons) {
+    btn.disabled = true;
   }
-
-  static addPrinter(newPrinter) {
-    // Insert Blank Row at top of printer list
-    if (
-      document.getElementById("printerNewTable").classList.contains("d-none")
-    ) {
-      document.getElementById("printerNewTable").classList.remove("d-none");
-    }
-
-    if (typeof newPrinter !== "undefined") {
-      document.getElementById("printerNewList").insertAdjacentHTML(
-        "beforebegin",
-        `
-   <tr id="newPrinterCard-${newPrintersIndex}">
-        <td><div class="mb-0">
-          <input id="newPrinterName-${newPrintersIndex}" type="text" class="form-control" placeholder="Leave blank to grab from OctoPrint" value="${newPrinter.name}">
-        </div></td>
-        <td><div class="mb-0">
-          <input id="newPrinterGroup-${newPrintersIndex}" type="text" class="form-control" placeholder="" value="${newPrinter.group}">
-        </div></td>
-        <td><div class="mb-0">
-          <input id="newPrinterURL-${newPrintersIndex}" type="text" class="form-control" placeholder="" value="${newPrinter.printerURL}">
-        </div></td>
-        <td><div class="mb-0">
-          <input id="newPrinterCamURL-${newPrintersIndex}" type="text" class="form-control" placeholder="Leave blank to grab from OctoPrint" value="${newPrinter.cameraURL}">
-        </div></td>
-        <td><div class="mb-0">
-          <input id="newPrinterAPIKEY-${newPrintersIndex}" type="text" class="form-control" placeholder="" value="${newPrinter.apikey}">
-        </div></td>
-        <td><button id="saveButton-${newPrintersIndex}" type="button" class="btn btn-success btn-sm">
-                <i class="fas fa-save"></i>
-            </button></td>
-        <td><button id="delButton-${newPrintersIndex}" type="button" class="btn btn-danger btn-sm">
-                <i class="fas fa-trash"></i>
-            </button></td>
-
-    </tr>
-  `
-      );
-    } else {
-      document.getElementById("printerNewList").insertAdjacentHTML(
-        "beforebegin",
-        `
-        <tr id="newPrinterCard-${newPrintersIndex}">
-        <td><div class="mb-0">
-          <input id="newPrinterName-${newPrintersIndex}" type="text" class="form-control" placeholder="Leave blank to grab from OctoPrint">
-          <small>Example: <code>My Awesome Printer Name</code></small>
-        </div></td>
-        <td><div class="mb-0">
-          <input id="newPrinterGroup-${newPrintersIndex}" type="text" class="form-control" placeholder="">
-          <small>Example: <code>Rack 1</code></small>
-        </div></td>
-        <td><div class="mb-0">
-          <input id="newPrinterURL-${newPrintersIndex}" type="text" class="form-control" placeholder="">
-          <small>Example: <code>http://192.168.1.5:80</code></small>
-        </div></td>
-        <td><div class="mb-0">
-          <input id="newPrinterCamURL-${newPrintersIndex}" type="text" class="form-control" placeholder="Leave blank to grab from OctoPrint">
-          <small>Example: <code>http://192.168.1.5/webcam/?action=stream</code></small>
-        </div></td>
-        <td><div class="mb-0">
-          <input id="newPrinterAPIKEY-${newPrintersIndex}" type="text" class="form-control" placeholder="">
-          <small>OctoPrint Version 1.4.1+: <code>Must use generated User/Application Key</code></small>
-        </div></td>
-        <td><button id="saveButton-${newPrintersIndex}" type="button" class="btn btn-success btn-sm">
-                <i class="fas fa-save"></i>
-            </button></td>
-        <td><button id="delButton-${newPrintersIndex}" type="button" class="btn btn-danger btn-sm">
-                <i class="fas fa-trash"></i>
-            </button></td>
-
-    </tr>
-  `
-      );
-    }
-    let currentIndex = JSON.parse(JSON.stringify(newPrintersIndex));
-    document
-      .getElementById(`delButton-${newPrintersIndex}`)
-      .addEventListener("click", (event) => {
-        removeLine(document.getElementById(`newPrinterCard-${currentIndex}`));
-        const table = document.getElementById("printerNewTable");
-        if (table.rows.length === 1) {
-          if (!table.classList.contains("d-none")) {
-            table.classList.add("d-none");
-          }
-        }
-      });
-    document
-      .getElementById(`saveButton-${newPrintersIndex}`)
-      .addEventListener("click", (event) => {
-        PrintersManagement.savePrinter(event.target);
-      });
-    const printerName = document.getElementById(
-      `newPrinterName-${newPrintersIndex}`
-    );
-    newPrintersIndex++;
+  for (const btn of onScreenDelete) {
+    btn.disabled = true;
   }
-
-  static async importPrinters() {
-    return function (e) {
-      const theBytes = e.target.result; // .split('base64,')[1];
-      // Initial JSON validation
-      if (Validate.JSON(theBytes)) {
-        // If we can parse the file.
-
-        // Grab uploaded file contents into an object
-        const importPrinters = JSON.parse(theBytes);
-        // Loop over import only importing printers with correct fields.
-        for (let index = 0; index < importPrinters.length; index++) {
-          const printer = {
-            printerURL: "Key not found",
-            cameraURL: "Key not found",
-            name: "Key not found",
-            group: "Key not found",
-            apikey: "Key not found"
-          };
-          if (typeof importPrinters[index].name !== "undefined") {
-            printer.name = importPrinters[index].name;
-          }
-          if (typeof importPrinters[index].printerURL !== "undefined") {
-            printer.printerURL = importPrinters[index].printerURL;
-          }
-          if (typeof importPrinters[index].cameraURL !== "undefined") {
-            printer.cameraURL = importPrinters[index].cameraURL;
-          }
-          if (typeof importPrinters[index].group !== "undefined") {
-            printer.group = importPrinters[index].group;
-          }
-          if (typeof importPrinters[index].apikey !== "undefined") {
-            printer.apikey = importPrinters[index].apikey;
-          }
-          PrintersManagement.addPrinter(printer);
-        }
-        UI.createAlert(
-          "success",
-          "Successfully imported your printer list, Please check it over and save when ready.",
-          3000
-        );
-      } else {
-        UI.createAlert(
-          "error",
-          "The file you have tried to upload contains json syntax errors.",
-          3000
-        );
-      }
-    };
+  UI.createAlert(
+    "warning",
+    "Starting to save all your instances... this may take some time...",
+    onScreenButtons.length * 1500
+  );
+  for (const btn of onScreenButtons) {
+    btn.disabled = false;
+    btn.click();
+    await delay(1500);
   }
+  UI.createAlert("success", "Successfully saved all your instances", 4000);
+  saveAllBtn.disabled = false;
+  deleteAllBtn.disabled = true;
+});
 
-  static async deletePrinter() {
-    if (deletedPrinters.length > 0) {
-      const post = await OctoFarmClient.post(
-        "printers/remove",
-        deletedPrinters
-      );
-      if (post.status === 200) {
-        let printersRemoved = await post.json();
-        printersRemoved = printersRemoved.printersRemoved;
-        printersRemoved.forEach((printer) => {
-          UI.createAlert(
-            "success",
-            `Printer: ${printer.printerURL} has successfully been removed from the farm...`,
-            1000,
-            "Clicked"
-          );
-          document.getElementById(`printerCard-${printer.printerId}`).remove();
-        });
-      } else {
-        UI.createAlert(
-          "error",
-          "Something went wrong updating the Server...",
-          3000,
-          "Clicked"
-        );
-      }
-    } else {
-      UI.createAlert(
-        "error",
-        "To delete a printer... one must first select a printer.",
-        3000,
-        "Clicked"
-      );
-    }
-  }
-
-  static async savePrinter(event) {
-    // Gather the printer data...
-    let newId = event.id.split("-");
-    newId = newId[1];
-
-    // Grab new printer cells...
-    const printerURL = document.getElementById(`newPrinterURL-${newId}`);
-    const printerCamURL = document.getElementById(`newPrinterCamURL-${newId}`);
-    const printerAPIKEY = document.getElementById(`newPrinterAPIKEY-${newId}`);
-    const printerGroup = document.getElementById(`newPrinterGroup-${newId}`);
-    const printerName = document.getElementById(`newPrinterName-${newId}`);
-
-    const errors = [];
-    let printCheck = -1;
-    if (printerURL.value !== "") {
-      printCheck = _.findIndex(printerInfo, function (o) {
-        return (
-          JSON.stringify(o.printerURL) === JSON.stringify(printerURL.value)
-        );
-      });
-    }
-    // Check information is filled correctly...
-    if (
-      printerURL.value === "" ||
-      printCheck > -1 ||
-      printerAPIKEY.value === "" ||
-      printerName.value === "" ||
-      printerCamURL.value === ""
-    ) {
-      if (printerURL.value === "") {
-        errors.push({ type: "warning", msg: "Please input your printers URL" });
-      }
-      if (printerAPIKEY.value === "") {
-        errors.push({
-          type: "warning",
-          msg: "Please input your printers API Key"
-        });
-      }
-      if (printCheck > -1) {
-        errors.push({
-          type: "error",
-          msg: `Printer URL: ${printerURL.value} already exists on farm`
-        });
-      }
-    }
-    if (errors.length > 0) {
-      errors.forEach((error) => {
-        UI.createAlert(error.type, error.msg, 3000, "clicked");
-      });
-    } else {
-      const printers = [];
-      const saveButton = document.getElementById(`saveButton-${newId}`);
-      saveButton.innerHTML = '<i class="fas fa-spinner fa-spin"></i>';
-      saveButton.disabled = true;
-      const printer = new PrintersManagement(
-        printerURL.value,
-        printerCamURL.value,
-        printerAPIKEY.value,
-        printerGroup.value,
-        printerName.value
-      ).build();
-      printers.push(printer);
-      const post = await OctoFarmClient.post("printers/add", printers);
-      if (post.status === 200) {
-        let printersAdded = await post.json();
-        printersAdded = printersAdded.printersAdded;
-        printersAdded.forEach((printer) => {
-          UI.createAlert(
-            "success",
-            `Printer: ${printer.printerURL} has successfully been added to the farm...`,
-            500,
-            "Clicked"
-          );
-        });
-        event.parentElement.parentElement.parentElement.remove();
-      } else {
-        UI.createAlert(
-          "error",
-          "Something went wrong updating the Server...",
-          3000,
-          "Clicked"
-        );
-        saveButton.innerHTML = '<i class="fas fa-save"></i>';
-        saveButton.disabled = false;
-      }
-    }
-    const table = document.getElementById("printerNewTable");
-    if (table.rows.length === 1) {
-      if (!table.classList.contains("d-none")) {
-        table.classList.add("d-none");
-      }
-    }
-  }
-
-  build() {
-    return this.printer;
-  }
-}
-
-class dashUpdate {}
-
+// Setup drag and drop re-ordering listeners
 const el = document.getElementById("printerList");
 const sortable = Sortable.create(el, {
   handle: ".sortableList",
